requirements.txt
# ============================================================================
# CropShield AI - Production Requirements
# ============================================================================
# Last Updated: November 11, 2025
# Python Version: 3.10+
# 
# Installation:
#   pip install -r requirements.txt
#
# For GPU support, install PyTorch with CUDA first:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# ============================================================================

# ============================================================================
# CORE DEEP LEARNING FRAMEWORK (REQUIRED)
# ============================================================================
torch>=2.0.0,<2.3.0              # PyTorch deep learning framework
torchvision>=0.15.0,<0.18.0      # Computer vision models and transforms
torchaudio>=2.0.0,<2.3.0         # Audio processing (optional, for completeness)

# ============================================================================
# IMAGE PROCESSING & COMPUTER VISION (REQUIRED)
# ============================================================================
Pillow>=10.0.0,<11.0.0           # Image processing library
opencv-python>=4.8.0,<5.0.0      # OpenCV for GradCAM and image operations
pillow-simd>=9.0.0               # SIMD-optimized Pillow (optional, faster)
albumentations>=1.3.0,<2.0.0     # Advanced data augmentation (optional)

# ============================================================================
# SCIENTIFIC COMPUTING & DATA PROCESSING (REQUIRED)
# ============================================================================
numpy>=1.24.0,<2.0.0             # Numerical computing
pandas>=2.0.0,<3.0.0             # Data manipulation and analysis
scikit-learn>=1.3.0,<2.0.0       # Machine learning utilities
scipy>=1.10.0,<2.0.0             # Scientific computing

# ============================================================================
# VISUALIZATION & PLOTTING (REQUIRED)
# ============================================================================
matplotlib>=3.7.0,<4.0.0         # Plotting and visualization
seaborn>=0.12.0,<1.0.0           # Statistical data visualization
plotly>=5.17.0,<6.0.0            # Interactive plots
altair>=5.0.0,<6.0.0             # Declarative visualization

# ============================================================================
# WEB APPLICATION FRAMEWORK (REQUIRED)
# ============================================================================
streamlit>=1.28.0,<2.0.0         # Web app framework
streamlit-extras>=0.3.0          # Additional Streamlit components

# ============================================================================
# PROGRESS BARS & CLI UTILITIES (REQUIRED)
# ============================================================================
tqdm>=4.66.0,<5.0.0              # Progress bars for loops
rich>=13.5.0,<14.0.0             # Rich text and formatting in terminal
colorama>=0.4.6,<1.0.0           # Cross-platform colored terminal text

# ============================================================================
# MODEL EXPORT & DEPLOYMENT (OPTIONAL)
# ============================================================================
onnx>=1.14.0,<2.0.0              # ONNX model format
onnxruntime>=1.15.0,<2.0.0       # ONNX inference runtime
onnxruntime-gpu>=1.15.0,<2.0.0   # ONNX GPU runtime (if CUDA available)

# ============================================================================
# DATA LOADING & SERIALIZATION (OPTIONAL)
# ============================================================================
webdataset>=0.2.0,<1.0.0         # Large-scale datasets (optional optimization)
h5py>=3.9.0,<4.0.0               # HDF5 file format support

# ============================================================================
# CONFIGURATION & UTILITIES (REQUIRED)
# ============================================================================
PyYAML>=6.0.0,<7.0.0             # YAML configuration files
python-dotenv>=1.0.0,<2.0.0      # Environment variable management
pathlib>=1.0.0                   # Path manipulation (built-in, for compatibility)

# ============================================================================
# TESTING & QUALITY ASSURANCE (DEVELOPMENT)
# ============================================================================
pytest>=7.4.0,<8.0.0             # Testing framework
pytest-cov>=4.1.0,<5.0.0         # Coverage reporting
black>=23.7.0,<24.0.0            # Code formatter
flake8>=6.1.0,<7.0.0             # Linting
mypy>=1.5.0,<2.0.0               # Type checking

# ============================================================================
# JUPYTER NOTEBOOK SUPPORT (DEVELOPMENT)
# ============================================================================
jupyter>=1.0.0,<2.0.0            # Jupyter notebook
ipykernel>=6.25.0,<7.0.0         # IPython kernel
ipywidgets>=8.1.0,<9.0.0         # Interactive widgets

# ============================================================================
# DOCUMENTATION GENERATION (DEVELOPMENT)
# ============================================================================
fpdf>=1.7.2,<2.0.0               # PDF generation (for reports)
markdown>=3.4.0,<4.0.0           # Markdown processing

# ============================================================================
# MISCELLANEOUS (OPTIONAL)
# ============================================================================
requests>=2.31.0,<3.0.0          # HTTP library
psutil>=5.9.0,<6.0.0             # System monitoring

